refactor(ir): Move verifier code to independent src/ir/verifier directory
Extract pure verifier files from src/ir/transforms/ into a dedicated
src/ir/verifier/ directory to improve code organization:

- Move verifier.cpp, property_verifier_registry.cpp, verify_ssa_pass.cpp,
  verify_no_nested_call_pass.cpp, type_check_pass.cpp to src/ir/verifier/
- Move verifier.h, property_verifier_registry.h, verification_error.h
  to include/pypto/ir/verifier/
- Update all include paths across transforms/ and bindings/
- Update CMakeLists.txt with new source paths under # IR - Verifier group
