+++ src/ir/transforms/flatten_single_stmt_pass.cpp
@@ -20,7 +20,7 @@
 #include "pypto/ir/transforms/pass_properties.h"
 #include "pypto/ir/transforms/passes.h"
 #include "pypto/ir/transforms/utils/flatten_single_stmt.h"
-#include "pypto/ir/transforms/verifier.h"
+#include "pypto/ir/verifier/verifier.h"
 
 namespace pypto {
 namespace ir {
