+++ src/ir/transforms/pass_context.cpp
@@ -19,8 +19,8 @@
 #include "pypto/core/logging.h"
 #include "pypto/ir/program.h"
 #include "pypto/ir/transforms/passes.h"
-#include "pypto/ir/transforms/property_verifier_registry.h"
-#include "pypto/ir/transforms/verifier.h"
+#include "pypto/ir/verifier/property_verifier_registry.h"
+#include "pypto/ir/verifier/verifier.h"
 
 namespace pypto {
 namespace ir {
