+++ src/ir/verifier/property_verifier_registry.cpp
@@ -9,7 +9,7 @@
  * -----------------------------------------------------------------------------------------------------------
  */
 
-#include "pypto/ir/transforms/property_verifier_registry.h"
+#include "pypto/ir/verifier/property_verifier_registry.h"
 
 #include <cstdint>
 #include <functional>
@@ -19,7 +19,7 @@
 #include "pypto/core/error.h"
 #include "pypto/ir/program.h"
 #include "pypto/ir/transforms/ir_property.h"
-#include "pypto/ir/transforms/verifier.h"
+#include "pypto/ir/verifier/verifier.h"
 
 namespace pypto {
 namespace ir {
